requirements.txt
pytest
playwright
allure-pytest
Pillow
numpy
